Remove RealmVersions and substitute with an Int counter.
