+++ examples/ios/swift/Migration/Examples/Example_v3.swift
@@ -25,12 +25,7 @@ import RealmSwift
 
 //// MARK: - Schema
 //
-//enum RealmVersion: Int, CaseIterable {
-//    case v0
-//    case v1
-//    case v2
-//    case v3
-//}
+//let schemaVersion = 3
 //
 //// Changes from previous version:
 //// rename the `Dog` object to `Pet`
